Replaced migration shell script by NodeJS script.
