+++ clients/typescript/package.json
@@ -44,7 +44,7 @@
     "dist"
   ],
   "bin": {
-    "electric": "dist/cli/bin.js"
+    "electric": "dist/cli/index.js"
   },
   "exports": {
     "./package.json": "./package.json",
@@ -100,8 +100,8 @@
   },
   "scripts": {
     "dev": "tsmodule dev",
-    "build": "rm -rf ./dist && tsc && tsmodule build && cp -R ./scripts/* ./dist",
-    "build-dev": "rm -rf ./dist && tsc && tsmodule build --dev && cp -R ./scripts/* ./dist",
+    "build": "rm -rf ./dist && tsc && tsmodule build",
+    "build-dev": "rm -rf ./dist && tsc && tsmodule build --dev",
     "test": "ava",
     "typecheck": "tsc -p tsconfig.tests.json",
     "posttest": "npm run typecheck",
@@ -138,11 +138,11 @@
     "frame-stream": "^3.0.1",
     "lodash.flow": "^3.5.0",
     "lodash.groupby": "^4.6.0",
+    "lodash.isequal": "^4.5.0",
     "lodash.mapvalues": "^4.6.0",
     "lodash.omitby": "^4.6.0",
     "lodash.partition": "^4.6.0",
     "lodash.pick": "^4.4.0",
-    "lodash.isequal": "^4.5.0",
     "lodash.throttle": "^4.1.1",
     "loglevel": "^1.8.1",
     "long": "^5.2.0",
@@ -164,6 +164,7 @@
     "@types/base-64": "^1.0.0",
     "@types/better-sqlite3": "7.6.3",
     "@types/cordova-sqlite-storage": "^1.5.5",
+    "@types/decompress": "^4.2.4",
     "@types/lodash.flow": "^3.5.7",
     "@types/lodash.groupby": "^4.6.7",
     "@types/lodash.isequal": "^4.5.6",
@@ -181,6 +182,7 @@
     "@typescript-eslint/parser": "^5.34.0",
     "ava": "^4.3.1",
     "better-sqlite3": "^8.4.0",
+    "decompress": "^4.2.1",
     "eslint": "^8.22.0",
     "expo-sqlite": "^11.0.0",
     "husky": "^8.0.3",
