+++ clients/typescript/test/migrators/builder.test.ts
@@ -1,6 +1,6 @@
 import test from 'ava'
 import { makeMigration, parseMetadata } from '../../src/migrators/builder'
-import { loadMigrations } from '../../src/cli'
+import { loadMigrations } from '../../src/cli/migrations/builder'
 import {
   SatOpMigrate,
   SatOpMigrate_Table,
